Desenvolvimento dos casos 1 e 3, otimizacao e simplificacao de codigo, criacao do ficheiro request_control.c

+++ p2/publisher/pub.c
@@ -1,62 +1,44 @@
-#include "utils/logging.h"
-#include "utils/fifo.h"
-#include <assert.h>
+#include "../utils/logging.h"
+#include "../utils/fifo.h"
+#include "../utils/request_control.h"
 #include <errno.h>
 #include <fcntl.h>
-#include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdint.h>
 #include <string.h>
-#include <sys/stat.h>
-#include <sys/types.h>
-#include <sys/wait.h>
 #include <unistd.h>
 
 #define BUFFER_SIZE 1024
 
 int main(int argc, char **argv) {
     (void)argv;
-
     if (argc != 4) {
         fprintf(stderr, "usage: pub <register_pipe_name> <pipe_name> <box_name>\n");
         return -1;
     }
 
-    // get initial arguments
-    char *register_pipe_name = argv[1];
-    char *pipe_name = argv[2];
-    char *box_name = argv[3];
+   if (send_request(1, argv[1], argv[2], argv[3]) == -1) return -1;
 
     // make pipe_name
+    char *pipe_name = argv[2];
     unlink(pipe_name);
     makefifo(pipe_name);
 
-    // open pipe to send request to mbroker
-    int fserv = open(register_pipe_name, O_WRONLY);
-    if (fserv < 0) return -1;
-    
-    // build request from publisher
-    char buffer[BUFFER_SIZE];
-    memset(buffer, 0, BUFFER_SIZE-1);
-    buffer[0] = (char)1;
-    memcpy(buffer+1, pipe_name, 256);
-    memcpy(buffer+1+256, box_name, 32);
-
-    ssize_t check;
-    // send request to mbroker
-    check = write(fserv, buffer, BUFFER_SIZE-1);
-    if (check < 0) return -1;
-    
     // open pipe to receive response from mbroker
     int fcli = open(pipe_name, O_RDONLY);
     if (fcli < 0) return -1;
-    
+
     // read response from mbroker
-    memset(buffer, 0, BUFFER_SIZE-1);
-    check = read(fcli, buffer, BUFFER_SIZE-1);
+    char buffer[BUFFER_SIZE];
+    ssize_t ret = read(fcli, buffer, BUFFER_SIZE-1);
+    if (ret < 0) return -1;
+
+    if (strcmp(buffer, "Connection refused") == 0) {
+        // (...)
+    } else {
+        // (...)
+    }
     
-    // do something with that...
     // close pipe to open it in O_WRONLY because it is a publisher
     close(fcli);
     if (strcmp(buffer, "-1") == 0) {
